requirements.txt
typer[all]==0.9.0
pyyaml==6.0.1
rich==13.7.0
